The newest libcw test finally tests something
The test was already doing something: it was getting data from a test run.
Now the test looks at the obtained data, compares it with reference data, and
giving some pass/fail results.

There is still some work to be done in the test, and some work to be done in
libcw (because the test actually indicates problems).