Automating a test of cw_gen_remove_last_character()
Validating correctness of the function with usage of easy rec.

Till now I was checking if the function under test was behaving correctly
only through what was played by generator.
Now the output produced by generator is also checked by a receiver. Receiver
receives played data and compares it with expected result.

One weak point of this test is that the generator is not perfect, and it may
receive something incorrectly.