BUGFIX for 'cw timing' problem
Problem:

Duration of inter-word-space was equal to 10 units, while it should have been
7 units.

As far as I know, the inter-word-space was not equal to 7 for many years.

Solution:

Add code that tracks enqueued spaces (primarily inter-mark-spaces and
inter-character-spaces, but also inter-word-spaces), and use that tracker to
decide how long an enqueued inter-word-space should really be.

Additional change that fixes the problem is with not enqueueing
inter-character-space when inter-word-space was enqueued.

Tests:

Large part of commits done to the repo since last release was dedicated to
expanding of testing capabilities of the libcw tests in area of generator,
receiver and timings.

Some libcw tests have been improved and other tests have been added to
automatically test some of libcw APIs that may be affected.

'simple receiver' type has been improved to make it more helpful in
discovering problems with timings.

Automatic test in which generated samples are stored to disc file and then
are analysed has been improved.

"Manual" tests have been executed where generated audio is captured by
external program (Audacity) and the different timings of different marks and
spaces are measured in the program to confirm that proper relations (1/3/7
units) are kept.