Moving and splitting src/cwutils/test_framework_tools.*
These new files that are a part of test framework have been moved to
src/test_framework/basic_utils and split into smaller modules.

The new files are compiled into a helper library that is linked to test code.

Hopefully in the future all the test-framework code (but not a code
implementing unit and integration tests) will go into src/test_framework.