Refactoring a logging function from test framework
The function is no longer a method of test executor - it is now a stand-alone
function. There was never a good reason to make this function (or any other
functions) the members of the executor.

The function is refactored and improved a bit.

The function is also renamed. New name is not as tied to "cw" package as the
old one.