Trying to detect 'unused' unit test functions
Adding some compiler flags and linker flags for compilation of libcw test
binary to try to detect situations where some test function is present in
code, but is not called.

With the additional flags the linker will remove such functions from binaries
and will log this to console. I may be able to notice such removal during
compilation.

This should help me detect situations where a test is defined, but by mistake
is not added to a set of executed tests.