Using per-sound-system receiver tolerances
There is no single tolerance configured on a receiver: each sound system will
get its specific tolerance.

I did only limited tests, only on one PC (so far). I have tested this code
with PulseAudio, and here 15% is OK.

I have also tested this code with ALSA, and it's definitely not ok. It was
equally bad in unmodified release package 3.6.0. The problem with ALSA is
either a bit older than 3.6.0 or is completely old, or is platform-specific.

I'll leave it like this for now.

OSS and console will have to be tested separately.

I still need to test Null sound system.