Using a better name of compilation flag for saving raw PCM samples
The previous commit did a change related to one of ./configure flags. This
commit is continuing changes in that area. Consistent naming of flags passed
to ./configure (both options start with '--enable-dev-') is being used. The
debug/dev feature enabled by this option is now clearly defined (not a bag of
'dev' features, but just a single feature related to raw PCM samples).