clang-tidy: "malloc" issues
Fixing some possible memory leaks reported by clang-tidy. Some of them were
located in combine_arguments(), and some of them in other function in
cwutils.

There is still one issue left in combine_arguments(), but unfortunately I
can't read anything useful from clang-tidy's report and find the situation in
which a leak would occur.

Therefore I'm leaving this one issue in code. I will try to fix it in the
future. Luckily this doesn't seem to be a serious issue because it exists in
code called by programs only once, and the leaked memory isn't large.