Marking 'cppcheck' and entire 'static code analysis' step as DONE
There are still numerous issues reported by cppcheck, but I'm leaving them
for later.

In particular there is an issue related to memory leak. It's in different
place than the one reported by clang-tidy, but as in case of that other issue
I think that it's not critical right now: it may occur only when realloc()
fails, it can occur only once (when the code is called from a program) and
the amount of memory is not large.

I will fix this issue in the future releases.