Renaming linting scripts from qa/ dir
Both scripts (one for clang-tidy and one for cppcheck) now have a name that
starts with "lint_".

In the future there may be more scripts that start some code linting tool, so
with common prefix in the name it will be easier to recognize which scripts
in the qa/ directory are related to linting.