Undo adding easy receiver to xcwcp
I decided to undo changes introduced to xcwcp at the beginning of this
development cycle, somewhere around of end of 2021 / beginning of 2022.

Easy receiver is not ready for prime time yet, and therefore can't be used in
xcwcp.

I will keep improving easy receiver and will re-introduce it in "production"
code sooner or later. Just not now.