PulseAudio tests: Debugging and working around a callback problem
The timings of generator when PulseAudio is used are rather poor. Wider
tolerance of receiver is a workaround for it (to make a test pass). Notice
that the tolerance set in this test (20%) is still lower than the default
tolerance of a receiver (50%).

Adding a 'ticket' to README.dev.org to investigate PulseAudio in the future.