Adding functions for getting random numbers to utility library
Will be used in next commit.

The functions are wrappers for libc functions. The wrappers allow me to
switch to different functions in the future. They also provide more
convenient API.

The wrapper for srand48() function also takes care of providing a good seed
to the wrapped function. We no longer use time(0), and this change will fix
one of TODOs.