Using new functions from 'random' module from cwutils/lib
The main place where the functions are used is test_framework: there was a
TODO that I wanted to address.

One of recent commits has been changing this area, so I thought that I can go
one step further and fix the TODO.

Notice that none of the changes related to random are really affecting
production code - just test code is affected.