Tests: adding program for automatic tests of cwgen
cwgen's only output is a string of words, so it's should be very easy to test
it automatically.

This commit is adding a test program that generates random options for cwgen,
runs cwgen with the options, inspects cwgen's output and decides if the
output is correct or not.

The test program is compiled and executed during "make check". The program is
integrated into build system using "non-recursive make" concept.