Tests: src/cwutils/tests are now built non-recursively
Similarly to recently added tests of cwgen, the tests in src/cwutils/tests
are now built using "non-recursive Makefile" concept.

One of outcomes is that results of the two tests (cwutils/tests and
cwgen/tests) are shown by build system in one common table/summary. The
person that ran "make check" can see the test results of unixcw in a more
readable way.

There are still two tests (one of them for libcw) that are done separately,
with a separate summary of results printed to console. In the future I would
like to re-make them in the same fashion (this will require only some changes
to build system).