Tests: header and footer of single libcw test are being updated
Layout of a header of libcw test has been slightly updated.

Adding a kite_on_test_completion() that prints a footer for a single libcw
test. Some header is already printed, but this function does it in a new way,
and it also increments failures/successes of tests executed by test executor.

The new kite_on_test_completion() function is used in two test functions as a
demonstration. The two functions were adjusted recently, so I'm adding one
more change to them.